refactor: member관련 클래스 네이밍과 폴더 위치 변경
