+++ src/main/java/com/moabam/global/error/model/ErrorMessage.java
@@ -7,7 +7,8 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public enum ErrorMessage {
 
-	INVALID_REQUEST_FIELD("올바른 요청 정보가 아닙니다.");
+	INVALID_REQUEST_FIELD("올바른 요청 정보가 아닙니다."),
+	LOGIN_FAILED("로그인에 실패했습니다.");
 
 	private final String message;
 }
